backend/requirements.txt
fastapi
uvicorn
langchain-openai
langgraph
python-dotenv
pydantic
langchain
langchain-groq
